Check that the template is a URL.
Previously, if a template was not a valid memegen template, it then went
on to check if the template existed as an image. But if the template
passed in was not a URL, this would fail with a 500. The bot will now
return a proper error message when requesting a template that is not
either a memegen template or an image URL.